requirements.txt
django>=1.9,<1.11
django-model-utils
jsonfield
six
netjsonconfig>=0.5.2,<0.6.0
django-sortedm2m>=1.3.2,<1.4
django-reversion>=2.0.6,<2.1
django-x509>=0.1.0,<0.2.0
djangorestframework
channels
asgi_redis
